+++ src/main/java/bank_bank/demo/repository/NasabahRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface NasabahRepository extends JpaRepository<Nasabah,Integer> {
+public interface NasabahRepository extends JpaRepository<Nasabah,Long> {
 
     List<Nasabah> findNasabahByIsDeletedFalse();
     Optional<Nasabah> findNasabahByNikAndIsDeletedFalse(String nik);
